修改重启条件

+++ watch.py
@@ -13,7 +13,8 @@ p = None
 
 class when_file_chanage(FileSystemEventHandler):
     def on_any_event(self, event):
-        p.send_signal(1)
+        if not event.is_directory and not event.src_path.endswith(('.pyc', '.swp', ".swx")):
+            p.send_signal(1)
 
 
 def HUP_handler(signum, frame):
